Removed seemingly extraneous line
The proof that d divides a relies on the remainder as given by the Division Theorem being 0 to satisfy the definition of "divides" given above. This does not require any assumptions or comparisons between d and a, and the note that d is less than or equal to a thus appears extraneous and distracts from the flow of the proof (as does the noting of the relatively trivial point that a is in the set X).